stm32_syscfg: convert to modern.
Includes the f4, l0 and wb.  f4 renames MEMRM to MEMRMP to be both
consistent with other parts and consistent with ref man.

Retested on the WB, but l0 and f4 code was simply moved.
For yaml files, given how varied syscfg is, we default to using the
family name as the type, but still allow overriding via explicit type in
the yaml file if desired.

Signed-off-by: Karl Palsson <karlp@tweak.net.au>